# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the workdadventure back office written at the Hochschule Bochum",
        "title": "Workadventure Admin Back Office API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "GitHub Issues",
            "url": "https://github.com/bo-mathadventure/admin"
        },
        "license": {
            "name": "AGPL 3.0",
            "url": "https://github.com/teamdigitale/licenses/blob/master/AGPL-3.0-or-later"
        },
        "version": "1.0"
    },
    "host": "localhost:4664",
    "basePath": "/",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Do a login with user credentials (email/password) when a password is set",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get Login Token",
                "parameters": [
                    {
                        "description": "-",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Start a registration of a new user. Only works when registration are enabled",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register new user",
                "parameters": [
                    {
                        "description": "-",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    }
                }
            }
        },
        "/auth/saml/acs": {
            "post": {
                "description": "Get SAML response of the IDP. This route is only available when SAML is correctly configured.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth",
                    "saml"
                ],
                "summary": "SAML Response Callback",
                "parameters": [
                    {
                        "type": "string",
                        "name": "RelayState",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "SAMLResponse",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    }
                }
            }
        },
        "/auth/saml/start": {
            "get": {
                "description": "Starts a new SAML authentication flow. This route is only available when SAML is correctly configured.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth",
                    "saml"
                ],
                "summary": "Get SAML Auth URL",
                "responses": {
                    "302": {
                        "description": "Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.APIResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean"
                },
                "extra": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "ERR_INVALID_PERMISSIONS"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "handler.LoginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "bob@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "my$ecur3P4$$word"
                }
            }
        },
        "handler.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "handler.RegisterRequest": {
            "type": "object",
            "properties": {
                "confirmPassword": {
                    "type": "string",
                    "example": "my$ecur3P4$$word"
                },
                "email": {
                    "type": "string",
                    "example": "bob@example.com"
                },
                "language": {
                    "type": "string",
                    "example": "de"
                },
                "password": {
                    "type": "string",
                    "example": "my$ecur3P4$$word"
                },
                "username": {
                    "type": "string",
                    "example": "Bob"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "JWT user token from login",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}